# services/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Global Probabilistic Weather Forecasting Platform API",
    "version": "1.1.0"
  },
  "paths": {
...
    "/v1/maps": {
      "get": {
        "summary": "Get map tile metadata",
        "description": "Return the tile template and legend for a weather map layer.\n\nUnder Lifecycle V2, supports either ``valid_time`` or ``lead_time_hours`` (with optional ``initial_time``).",
        "operationId": "get_spatial_layer_v1_maps_get",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A model identifier.",
              "title": "Model"
            },
            "description": "A model identifier."
          },
          {
            "name": "variable",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A forecast variable code.",
              "title": "Variable"
            },
            "description": "A forecast variable code."
          },
          {
            "name": "level",
            "in": "query",
            "required": true,
            "schema": {
              "const": "surface",
              "type": "string",
              "description": "The vertical level; only 'surface' is supported.",
              "title": "Level"
            },
            "description": "The vertical level; only 'surface' is supported."
          },
          {
            "name": "lead_time_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Forecast offset hours from cycle time.",
              "title": "Lead Time Hours"
            },
            "description": "Forecast offset hours from cycle time."
          },
          {
            "name": "valid_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional ISO 8601 UTC valid time (Lifecycle V2). Resolves to the newest serveable cycle capable of providing this valid time.",
              "title": "Valid Time"
            },
            "description": "Optional ISO 8601 UTC valid time (Lifecycle V2). Resolves to the newest serveable cycle capable of providing this valid time."
          },
          {
            "name": "initial_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional ISO 8601 UTC cycle time pinning the model run (e.g. 2026-08-13T00:00:00Z).",
              "title": "Initial Time"
            },
            "description": "Optional ISO 8601 UTC cycle time pinning the model run (e.g. 2026-08-13T00:00:00Z)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpatialLayerEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maps/{model}/wind_10m/vector-field": {
      "get": {
        "summary": "Get 10-meter wind vector field for flow animation",
        "description": "Return the quantized Int16 binary wind vector field for particle flow animation.",
        "operationId": "get_wind_vector_field_v1_maps__model__wind_10m_vector_field_get",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          },
          {
            "name": "lead_time_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Forecast offset hours from cycle time.",
              "title": "Lead Time Hours"
            },
            "description": "Forecast offset hours from cycle time."
          },
          {
            "name": "valid_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional ISO 8601 UTC valid time (Lifecycle V2).",
              "title": "Valid Time"
            },
            "description": "Optional ISO 8601 UTC valid time (Lifecycle V2)."
          },
          {
            "name": "initial_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional ISO 8601 UTC cycle time pinning the model run.",
              "title": "Initial Time"
            },
            "description": "Optional ISO 8601 UTC cycle time pinning the model run."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/maps/{model}/{variable}/{level}/{z}/{x}/{y}.png": {
      "get": {
        "summary": "Render a forecast raster tile",
        "description": "Render a 256x256 PNG tile of the forecast field for the selection.",
        "operationId": "get_map_tile_v1_maps__model___variable___level___z___x___y__png_get",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          },
          {
            "name": "variable",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variable"
            }
          },
          {
            "name": "level",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Level"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Z"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Y"
            }
          },
          {
            "name": "lead_time_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Forecast offset hours from cycle time.",
              "title": "Lead Time Hours"
            },
            "description": "Forecast offset hours from cycle time."
          },
          {
            "name": "valid_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional ISO 8601 UTC valid time (Lifecycle V2).",
              "title": "Valid Time"
            },
            "description": "Optional ISO 8601 UTC valid time (Lifecycle V2)."
          },
          {
            "name": "initial_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional ISO 8601 UTC cycle time pinning the model run.",
              "title": "Initial Time"
            },
            "description": "Optional ISO 8601 UTC cycle time pinning the model run."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpatialLayerData": {
        "properties": {
          "tile_url_template": {
            "type": "string",
            "title": "Tile Url Template"
          },
          "min_zoom": {
            "type": "integer",
            "title": "Min Zoom"
          },
          "max_zoom": {
            "type": "integer",
            "title": "Max Zoom"
          },
          "lead_time_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Hours"
          },
          "valid_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Time"
          },
          "legend": {
            "$ref": "#/components/schemas/SpatialLayerLegend"
          },
          "vector_field_url_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Field Url Template"
          }
        },
        "type": "object",
        "required": ["tile_url_template", "min_zoom", "max_zoom", "legend"],
        "title": "SpatialLayerData",
        "description": "The payload of a spatial layer (API.md section 4.1)."
      },
      "SpatialLayerEnvelope": {
        "properties": {
          "object": {
            "type": "string",
            "const": "spatial_layer",
            "title": "Object",
            "default": "spatial_layer"
          },
          "data": {
            "$ref": "#/components/schemas/SpatialLayerData"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": ["data"],
        "title": "SpatialLayerEnvelope",
        "description": "The spatial layer response envelope (API.md sections 4.1 and 2.3)."
      },
      "SpatialLayerLegend": {
        "properties": {
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "stops": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "string"
                  }
                ]
              },
              "type": "array"
            },
            "type": "array",
            "title": "Stops"
          }
        },
        "type": "object",
        "required": ["unit", "stops"],
        "title": "SpatialLayerLegend",
        "description": "The legend of a spatial layer (API.md section 4.1)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}